… docs/feature_rgb_matrix.md
… lib/lib8tion/lib8tion.h
… lib/lib8tion/random8.h
… lib/lib8tion/scale8.h
… lib/lib8tion/trig8.h
… quantum/rgb_matrix.c
… quantum/rgb_matrix_animations/alpha_mods_anim.h
… quantum/rgb_matrix_animations/breathing_anim.h
… quantum/rgb_matrix_animations/cycle_all_anim.h
… quantum/rgb_matrix_animations/cycle_left_right_anim.h
… quantum/rgb_matrix_animations/cycle_up_down_anim.h
… quantum/rgb_matrix_animations/digital_rain_anim.h
… quantum/rgb_matrix_animations/dual_beacon_anim.h
… quantum/rgb_matrix_animations/gradient_up_down_anim.h
… quantum/rgb_matrix_animations/jellybean_raindrops_anim.h
… quantum/rgb_matrix_animations/rainbow_beacon_anim.h
… quantum/rgb_matrix_animations/rainbow_moving_chevron_anim.h
… quantum/rgb_matrix_animations/rainbow_pinwheels_anim.h
… quantum/rgb_matrix_animations/raindrops_anim.h
… quantum/rgb_matrix_animations/solid_color_anim.h
… quantum/rgb_matrix_animations/solid_reactive_anim.h
… quantum/rgb_matrix_animations/solid_reactive_simple_anim.h
… quantum/rgb_matrix_animations/solid_splash_anim.h
… quantum/rgb_matrix_animations/splash_anim.h
… quantum/rgb_matrix_types.h