@XScorpion2 2019-03-17
Ready for review as this is about as performant as I can get it without getting to crazy. This should be merged after https://github.com/qmk/qmk_firmware/pull/5328 and https://github.com/qmk/qmk_firmware/pull/5330 so I can fix any merge errors that will occur from this pr when they land. I've already manually merged those into another test branch and verified they will work with my changes here with a few minor adjustments.