@XScorpion2 2019-03-17
https://github.com/qmk/qmk_firmware/pull/5330 was merged to master, master has been merged into here and the new matrix anim effect has been updated to match this refactor work.