@drashna 2019-03-20
At this point, I'm not sure if anything else needs to be changed/added/improved here. At least code wise.

However, could you update the documentation for the RGB Matrix? 

There is at least one new setting that has been added, and the effects order has been changed. 
If you could update the docs to reflect these changes and additions, that would be awesome!