@patrickmt 2019-03-20
I benchmarked all these animations on a CTRL keyboard using @XScorpion2's dev-rgb_ctrl branch and RGB_MATRIX_LED_PROCESS_LIMIT set to 15 to mimic the Massdrop lighting configuration. The CTRL has 119 RGB LEDs and runs at 40MHz. This means that after 8 calls of rgb_matrix_task, all LEDs are processed and ready to output.

Most effects work very well, taking around 1.9ms to 2.6ms for animations to complete. Note that these times include any other code running between effect processing, including hardware ADC sampling tasks. Actual effect only processing times are more around 1.2ms to 1.9ms. The few issues I saw were with digital rain and multisplash effects where those took around 6ms. These tests were performed on the branch at 53053ec.

Because of various time sensitive tasks taking place on the CTRL, breaking up the LED processing has helped tremendously in making it operate as intended while using the RGB Matrix effects. One thing missing still is the ability to control the rate of flushes instead of flushing as fast as possible. This would allow one to target an update rate while giving the system time to operate without possibly expensive LED processing interfering.

I've discussed all this with @XScorpion2 and he said he will add in the flush rate timing ability.