@XScorpion2 2019-03-21
Still want to document each of the animations, but other than that I think it is good to go.