@patrickmt 2019-03-21
This PR is working very nice now, tested on both CTRL and ALT keyboards!