@XScorpion2 2019-03-26
Minor api change for future compatibility of things like this:
https://photos.app.goo.gl/waCkRJwsqjEBiFCCA