@XScorpion2 2019-03-27
Fixed a compile error that would occur if you had just `RGB_MATRIX_KEYRELEASES` defined