@drashna 2019-04-01
I noticed some issues with resuming from sleep on my board. I'm not sure if it's related to these changes, though.  Will test and let you know.