@drashna 2019-04-02
Lockup was a red herring.  It looks like it maybe some other change, and not this PR