@drashna 2019-04-02
> @drashna tease

Maybe, but I'd rather say something and have testing prove me wrong, then not say anything and have a bug merged in